Cleaned code, Master is now a "template"
Master branch is used as a template base. For example, at the start of the 2017 season a branch named "dev/2017" is branched from master. The new "dev/2017" branch now has all the base code ready for coding. Repeated code, such as joysticks, can be added to the master branch so that when "dev/year" is branched, it already has all those such parts.